rename layer preview files to comply with new naming standard
