+++ components/admin/data/layers/form/layer-preview/reducer.js
